chore(refactor): replace path
